% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4u-package.R
\docType{package}
\name{lme4u-package}
\alias{lme4u}
\alias{lme4u-package}
\title{lme4u: Interpretation and Diagnostics for 'lme4' Hierarchical Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interpret and diagnose linear mixed-effects models fitted with 'lme4'. Provides tools for summarizing model output, visualizing assumptions checks, and performing hypothesis tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hollyyfc/lme4u}
  \item \url{https://hollyyfc.github.io/lme4u/}
  \item Report bugs at \url{https://github.com/hollyyfc/lme4u/issues}
}

}
\author{
\strong{Maintainer}: Holly Cui \email{hollycui1218@gmail.com}

}
\keyword{internal}
