% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_lmer.R
\name{explain_lmer}
\alias{explain_lmer}
\title{Interpret Linear Mixed-Effects Model}
\usage{
explain_lmer(model, details = "general")
}
\arguments{
\item{model}{A fitted model object from \code{\link[lme4:lmer]{lme4::lmer()}}.}

\item{details}{Character. Specify which details to return: \code{"general"} (the default),
\code{"fixed"}, \code{"random"}, or \code{"full"}.}
}
\value{
A character string with the interpretation of the model.
}
\description{
Provides a user-friendly interpretation of a fitted lmer model's fixed and
random effects.
}
\examples{
library(lme4)
model <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
explain_lmer(model)
}
